Trying to add sprinting functionality
Can someone make the triceratop’s “panic” speed faster? Currently, the
triceratops doesn’t move fast enough to realistically require a
sprinting animation. Due to the fluctuating nature of the f1 parameter,
the animation might bleed over into the walk unless the value at which
it is triggered is raised.